+++ test/lib/useOldTz.js
@@ -6,6 +6,7 @@ const changeTz = (s) => {
   timezones['pacific/fiji'].dst = '01/15:02->11/05:03'
   timezones['europe/brussels'].dst = '03/29:02->10/25:03'
   timezones['america/chicago'].dst = '03/08:02->11/01:02'
+  timezones['canada/pacific'].dst = '03/08:02->11/01:02'
   s.timezones = timezones
   return s
 }