+++ zonefile/iana.js
@@ -58,12 +58,12 @@ module.exports = {
   'africa/casablanca': {
     offset: 1,
     hem: 'n',
-    dst: '04/19:03->05/31:02'
+    dst: '04/11:03->05/16:02'
   },
   'africa/ceuta': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'africa/conakry': {
     offset: 0,
@@ -88,7 +88,7 @@ module.exports = {
   'africa/el_aaiun': {
     offset: 1,
     hem: 'n',
-    dst: '04/19:03->05/31:02'
+    dst: '04/11:03->05/16:02'
   },
   'africa/freetown': {
     offset: 0,
@@ -222,12 +222,12 @@ module.exports = {
   'america/adak': {
     offset: -9,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/anchorage': {
     offset: -8,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/anguilla': {
     offset: -4,
@@ -252,7 +252,7 @@ module.exports = {
   'america/asuncion': {
     offset: -4,
     hem: 's',
-    dst: '03/21:24->10/04:00'
+    dst: '03/27:24->10/03:00'
   },
   'america/atikokan': {
     offset: -5,
@@ -261,7 +261,7 @@ module.exports = {
   'america/atka': {
     offset: -9,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/bahia': {
     offset: -3,
@@ -270,7 +270,7 @@ module.exports = {
   'america/bahia_banderas': {
     offset: -5,
     hem: 'n',
-    dst: '04/05:02->10/25:02'
+    dst: '04/04:02->10/31:02'
   },
   'america/barbados': {
     offset: -4,
@@ -299,7 +299,7 @@ module.exports = {
   'america/boise': {
     offset: -6,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/buenos_aires': {
     offset: -3,
@@ -308,7 +308,7 @@ module.exports = {
   'america/cambridge_bay': {
     offset: -6,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/campo_grande': {
     offset: -4,
@@ -338,12 +338,12 @@ module.exports = {
   'america/chicago': {
     offset: -5,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/chihuahua': {
     offset: -6,
     hem: 'n',
-    dst: '04/05:02->10/25:02'
+    dst: '04/04:02->10/31:02'
   },
   'america/coral_harbour': {
     offset: -5,
@@ -386,12 +386,12 @@ module.exports = {
   'america/denver': {
     offset: -6,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/detroit': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/dominica': {
     offset: -4,
@@ -400,7 +400,7 @@ module.exports = {
   'america/edmonton': {
     offset: -6,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/eirunepe': {
     offset: -5,
@@ -413,7 +413,7 @@ module.exports = {
   'america/ensenada': {
     offset: -7,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/fort_nelson': {
     offset: -7,
@@ -422,7 +422,7 @@ module.exports = {
   'america/fort_wayne': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/fortaleza': {
     offset: -3,
@@ -431,22 +431,22 @@ module.exports = {
   'america/glace_bay': {
     offset: -3,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/godthab': {
     offset: -2,
     hem: 'n',
-    dst: '03/28:22->10/24:23'
+    dst: '03/27:22->10/30:23'
   },
   'america/goose_bay': {
     offset: -3,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/grand_turk': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/grenada': {
     offset: -4,
@@ -471,12 +471,12 @@ module.exports = {
   'america/halifax': {
     offset: -3,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/havana': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:00->11/01:01'
+    dst: '03/14:00->11/07:01'
   },
   'america/hermosillo': {
     offset: -7,
@@ -490,17 +490,17 @@ module.exports = {
   'america/indianapolis': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/inuvik': {
     offset: -6,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/iqaluit': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/jamaica': {
     offset: -5,
@@ -513,7 +513,7 @@ module.exports = {
   'america/juneau': {
     offset: -8,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/kentucky': {
     offset: -4,
@@ -523,7 +523,7 @@ module.exports = {
   'america/knox_in': {
     offset: -5,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/kralendijk': {
     offset: -4,
@@ -540,12 +540,12 @@ module.exports = {
   'america/los_angeles': {
     offset: -7,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/louisville': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/lower_princes': {
     offset: -4,
@@ -574,12 +574,12 @@ module.exports = {
   'america/matamoros': {
     offset: -5,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/mazatlan': {
     offset: -6,
     hem: 'n',
-    dst: '04/05:02->10/25:02'
+    dst: '04/04:02->10/31:02'
   },
   'america/mendoza': {
     offset: -3,
@@ -588,37 +588,37 @@ module.exports = {
   'america/menominee': {
     offset: -5,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/merida': {
     offset: -5,
     hem: 'n',
-    dst: '04/05:02->10/25:02'
+    dst: '04/04:02->10/31:02'
   },
   'america/metlakatla': {
     offset: -8,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/mexico_city': {
     offset: -5,
     hem: 'n',
-    dst: '04/05:02->10/25:02'
+    dst: '04/04:02->10/31:02'
   },
   'america/miquelon': {
     offset: -2,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/moncton': {
     offset: -3,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/monterrey': {
     offset: -5,
     hem: 'n',
-    dst: '04/05:02->10/25:02'
+    dst: '04/04:02->10/31:02'
   },
   'america/montevideo': {
     offset: -3,
@@ -627,7 +627,7 @@ module.exports = {
   'america/montreal': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/montserrat': {
     offset: -4,
@@ -636,22 +636,22 @@ module.exports = {
   'america/nassau': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/new_york': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/nipigon': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/nome': {
     offset: -8,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/noronha': {
     offset: -2,
@@ -660,7 +660,7 @@ module.exports = {
   'america/nuuk': {
     offset: -3,
     hem: 'n',
-    dst: '03/28:22->10/24:23'
+    dst: '03/27:22->10/30:23'
   },
   'america/north_dakota': {
     offset: -5,
@@ -670,7 +670,7 @@ module.exports = {
   'america/ojinaga': {
     offset: -6,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/panama': {
     offset: -5,
@@ -679,7 +679,7 @@ module.exports = {
   'america/pangnirtung': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/paramaribo': {
     offset: -3,
@@ -692,7 +692,7 @@ module.exports = {
   'america/port-au-prince': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/port_of_spain': {
     offset: -4,
@@ -717,12 +717,12 @@ module.exports = {
   'america/rainy_river': {
     offset: -5,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/rankin_inlet': {
     offset: -5,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/recife': {
     offset: -3,
@@ -735,7 +735,7 @@ module.exports = {
   'america/resolute': {
     offset: -5,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/rio_branco': {
     offset: -5,
@@ -748,7 +748,7 @@ module.exports = {
   'america/santa_isabel': {
     offset: -7,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/santarem': {
     offset: -3,
@@ -757,7 +757,7 @@ module.exports = {
   'america/santiago': {
     offset: -4,
     hem: 's',
-    dst: '04/04:24->09/06:00'
+    dst: '04/03:24->09/05:00'
   },
   'america/santo_domingo': {
     offset: -4,
@@ -770,17 +770,17 @@ module.exports = {
   'america/scoresbysund': {
     offset: 0,
     hem: 'n',
-    dst: '03/29:00->10/25:01'
+    dst: '03/28:00->10/31:01'
   },
   'america/shiprock': {
     offset: -6,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/sitka': {
     offset: -8,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/st_barthelemy': {
     offset: -4,
@@ -789,7 +789,7 @@ module.exports = {
   'america/st_johns': {
     offset: -2.5,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/st_kitts': {
     offset: -4,
@@ -818,22 +818,22 @@ module.exports = {
   'america/thule': {
     offset: -3,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/thunder_bay': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/tijuana': {
     offset: -7,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/toronto': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/tortola': {
     offset: -4,
@@ -842,7 +842,7 @@ module.exports = {
   'america/vancouver': {
     offset: -7,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/virgin': {
     offset: -4,
@@ -856,17 +856,17 @@ module.exports = {
   'america/winnipeg': {
     offset: -5,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/yakutat': {
     offset: -8,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'america/yellowknife': {
     offset: -6,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'antarctica/casey': {
     offset: 8,
@@ -884,7 +884,7 @@ module.exports = {
   'antarctica/macquarie': {
     offset: 11,
     hem: 's',
-    dst: '04/05:03->10/04:02'
+    dst: '04/04:03->10/03:02'
   },
   'antarctica/mawson': {
     offset: 5,
@@ -893,7 +893,7 @@ module.exports = {
   'antarctica/mcmurdo': {
     offset: 12,
     hem: 's',
-    dst: '04/05:03->09/27:02'
+    dst: '04/04:03->09/26:02'
   },
   'antarctica/palmer': {
     offset: -4,
@@ -907,7 +907,7 @@ module.exports = {
   'antarctica/south_pole': {
     offset: 12,
     hem: 's',
-    dst: '04/05:03->09/27:02'
+    dst: '04/04:03->09/26:02'
   },
   'antarctica/syowa': {
     offset: 3,
@@ -916,7 +916,7 @@ module.exports = {
   'antarctica/troll': {
     offset: 2,
     hem: 's',
-    dst: '03/29:02->10/25:02'
+    dst: '03/28:02->10/31:02'
   },
   'antarctica/vostok': {
     offset: 6,
@@ -925,7 +925,7 @@ module.exports = {
   'arctic/longyearbyen': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'asia/aden': {
     offset: 3,
@@ -938,7 +938,7 @@ module.exports = {
   'asia/amman': {
     offset: 3,
     hem: 'n',
-    dst: '03/27:00->10/30:01'
+    dst: '03/26:00->10/29:01'
   },
   'asia/anadyr': {
     offset: 12,
@@ -987,7 +987,7 @@ module.exports = {
   'asia/beirut': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:00->10/24:24'
+    dst: '03/28:00->10/30:24'
   },
   'asia/bishkek': {
     offset: 6,
@@ -1028,7 +1028,7 @@ module.exports = {
   'asia/damascus': {
     offset: 3,
     hem: 'n',
-    dst: '03/27:00->10/29:24'
+    dst: '03/26:00->10/28:24'
   },
   'asia/dhaka': {
     offset: 6,
@@ -1049,12 +1049,12 @@ module.exports = {
   'asia/famagusta': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:03->10/25:04'
+    dst: '03/28:03->10/31:04'
   },
   'asia/gaza': {
     offset: 3,
     hem: 'n',
-    dst: '03/28:00->10/24:01'
+    dst: '03/27:00->10/30:01'
   },
   'asia/harbin': {
     offset: 8,
@@ -1063,7 +1063,7 @@ module.exports = {
   'asia/hebron': {
     offset: 3,
     hem: 'n',
-    dst: '03/28:00->10/24:01'
+    dst: '03/27:00->10/30:01'
   },
   'asia/ho_chi_minh': {
     offset: 7,
@@ -1096,7 +1096,7 @@ module.exports = {
   'asia/jerusalem': {
     offset: 3,
     hem: 'n',
-    dst: '03/27:02->10/25:02'
+    dst: '03/26:02->10/31:02'
   },
   'asia/kabul': {
     offset: 4.5,
@@ -1173,7 +1173,7 @@ module.exports = {
   'asia/nicosia': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:03->10/25:04'
+    dst: '03/28:03->10/31:04'
   },
   'asia/novokuznetsk': {
     offset: 7,
@@ -1270,12 +1270,12 @@ module.exports = {
   'asia/tehran': {
     offset: 4.5,
     hem: 'n',
-    dst: '03/21:00->09/20:24'
+    dst: '03/22:00->09/21:24'
   },
   'asia/tel_aviv': {
     offset: 3,
     hem: 'n',
-    dst: '03/27:02->10/25:02'
+    dst: '03/26:02->10/31:02'
   },
   'asia/thimbu': {
     offset: 6,
@@ -1337,17 +1337,17 @@ module.exports = {
   'atlantic/azores': {
     offset: 0,
     hem: 'n',
-    dst: '03/29:00->10/25:01'
+    dst: '03/28:00->10/31:01'
   },
   'atlantic/bermuda': {
     offset: -3,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'atlantic/canary': {
     offset: 1,
     hem: 'n',
-    dst: '03/29:01->10/25:02'
+    dst: '03/28:01->10/31:02'
   },
   'atlantic/cape_verde': {
     offset: -1,
@@ -1356,22 +1356,22 @@ module.exports = {
   'atlantic/faeroe': {
     offset: 1,
     hem: 'n',
-    dst: '03/29:01->10/25:02'
+    dst: '03/28:01->10/31:02'
   },
   'atlantic/faroe': {
     offset: 1,
     hem: 'n',
-    dst: '03/29:01->10/25:02'
+    dst: '03/28:01->10/31:02'
   },
   'atlantic/jan_mayen': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'atlantic/madeira': {
     offset: 1,
     hem: 'n',
-    dst: '03/29:01->10/25:02'
+    dst: '03/28:01->10/31:02'
   },
   'atlantic/reykjavik': {
     offset: 0,
@@ -1392,12 +1392,12 @@ module.exports = {
   'australia/act': {
     offset: 10,
     hem: 's',
-    dst: '04/05:03->10/04:02'
+    dst: '04/04:03->10/03:02'
   },
   'australia/adelaide': {
     offset: 9.5,
     hem: 's',
-    dst: '04/05:03->10/04:02'
+    dst: '04/04:03->10/03:02'
   },
   'australia/brisbane': {
     offset: 10,
@@ -1406,17 +1406,17 @@ module.exports = {
   'australia/broken_hill': {
     offset: 9.5,
     hem: 's',
-    dst: '04/05:03->10/04:02'
+    dst: '04/04:03->10/03:02'
   },
   'australia/canberra': {
     offset: 10,
     hem: 's',
-    dst: '04/05:03->10/04:02'
+    dst: '04/04:03->10/03:02'
   },
   'australia/currie': {
     offset: 10,
     hem: 's',
-    dst: '04/05:03->10/04:02'
+    dst: '04/04:03->10/03:02'
   },
   'australia/darwin': {
     offset: 9.5,
@@ -1429,12 +1429,12 @@ module.exports = {
   'australia/hobart': {
     offset: 10,
     hem: 's',
-    dst: '04/05:03->10/04:02'
+    dst: '04/04:03->10/03:02'
   },
   'australia/lhi': {
     offset: 10.5,
     hem: 's',
-    dst: '04/05:01->10/04:02'
+    dst: '04/04:01->10/03:02'
   },
   'australia/lindeman': {
     offset: 10,
@@ -1443,17 +1443,17 @@ module.exports = {
   'australia/lord_howe': {
     offset: 10.5,
     hem: 's',
-    dst: '04/05:01->10/04:02'
+    dst: '04/04:01->10/03:02'
   },
   'australia/melbourne': {
     offset: 10,
     hem: 's',
-    dst: '04/05:03->10/04:02'
+    dst: '04/04:03->10/03:02'
   },
   'australia/nsw': {
     offset: 10,
     hem: 's',
-    dst: '04/05:03->10/04:02'
+    dst: '04/04:03->10/03:02'
   },
   'australia/north': {
     offset: 9.5,
@@ -1470,22 +1470,22 @@ module.exports = {
   'australia/south': {
     offset: 9.5,
     hem: 's',
-    dst: '04/05:03->10/04:02'
+    dst: '04/04:03->10/03:02'
   },
   'australia/sydney': {
     offset: 10,
     hem: 's',
-    dst: '04/05:03->10/04:02'
+    dst: '04/04:03->10/03:02'
   },
   'australia/tasmania': {
     offset: 10,
     hem: 's',
-    dst: '04/05:03->10/04:02'
+    dst: '04/04:03->10/03:02'
   },
   'australia/victoria': {
     offset: 10,
     hem: 's',
-    dst: '04/05:03->10/04:02'
+    dst: '04/04:03->10/03:02'
   },
   'australia/west': {
     offset: 8,
@@ -1494,7 +1494,7 @@ module.exports = {
   'australia/yancowinna': {
     offset: 9.5,
     hem: 's',
-    dst: '04/05:03->10/04:02'
+    dst: '04/04:03->10/03:02'
   },
   'brazil/acre': {
     offset: -5,
@@ -1515,12 +1515,12 @@ module.exports = {
   'canada/atlantic': {
     offset: -3,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'canada/central': {
     offset: -5,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'canada/east-saskatchewan': {
     offset: -6,
@@ -1529,22 +1529,22 @@ module.exports = {
   'canada/eastern': {
     offset: -4,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'canada/mountain': {
     offset: -6,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'canada/newfoundland': {
     offset: -2.5,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'canada/pacific': {
     offset: -7,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'canada/saskatchewan': {
     offset: -6,
@@ -1558,22 +1558,22 @@ module.exports = {
   'chile/continental': {
     offset: -4,
     hem: 's',
-    dst: '04/04:24->09/06:00'
+    dst: '04/03:24->09/05:00'
   },
   'chile/easterisland': {
     offset: -6,
     hem: 's',
-    dst: '04/04:22->09/05:22'
+    dst: '04/03:22->09/04:22'
   },
   'europe/amsterdam': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/andorra': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/astrakhan': {
     offset: 4,
@@ -1582,82 +1582,82 @@ module.exports = {
   'europe/athens': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:03->10/25:04'
+    dst: '03/28:03->10/31:04'
   },
   'europe/belfast': {
     offset: 1,
     hem: 'n',
-    dst: '03/29:01->10/25:02'
+    dst: '03/28:01->10/31:02'
   },
   'europe/belgrade': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/berlin': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/bratislava': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/brussels': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/bucharest': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:03->10/25:04'
+    dst: '03/28:03->10/31:04'
   },
   'europe/budapest': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/busingen': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/chisinau': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/copenhagen': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/dublin': {
     offset: 1,
     hem: 'n',
-    dst: '03/29:01->10/25:02'
+    dst: '03/28:01->10/31:02'
   },
   'europe/gibraltar': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/guernsey': {
     offset: 1,
     hem: 'n',
-    dst: '03/29:01->10/25:02'
+    dst: '03/28:01->10/31:02'
   },
   'europe/helsinki': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:03->10/25:04'
+    dst: '03/28:03->10/31:04'
   },
   'europe/isle_of_man': {
     offset: 1,
     hem: 'n',
-    dst: '03/29:01->10/25:02'
+    dst: '03/28:01->10/31:02'
   },
   'europe/istanbul': {
     offset: 3,
@@ -1666,7 +1666,7 @@ module.exports = {
   'europe/jersey': {
     offset: 1,
     hem: 'n',
-    dst: '03/29:01->10/25:02'
+    dst: '03/28:01->10/31:02'
   },
   'europe/kaliningrad': {
     offset: 2,
@@ -1679,42 +1679,42 @@ module.exports = {
   'europe/kiev': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:03->10/25:04'
+    dst: '03/28:03->10/31:04'
   },
   'europe/lisbon': {
     offset: 1,
     hem: 'n',
-    dst: '03/29:01->10/25:02'
+    dst: '03/28:01->10/31:02'
   },
   'europe/ljubljana': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/london': {
     offset: 1,
     hem: 'n',
-    dst: '03/29:01->10/25:02'
+    dst: '03/28:01->10/31:02'
   },
   'europe/luxembourg': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/madrid': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/malta': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/mariehamn': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:03->10/25:04'
+    dst: '03/28:03->10/31:04'
   },
   'europe/minsk': {
     offset: 3,
@@ -1723,7 +1723,7 @@ module.exports = {
   'europe/monaco': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/moscow': {
     offset: 3,
@@ -1732,37 +1732,37 @@ module.exports = {
   'europe/nicosia': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:03->10/25:04'
+    dst: '03/28:03->10/31:04'
   },
   'europe/oslo': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/paris': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/podgorica': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/prague': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/riga': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:03->10/25:04'
+    dst: '03/28:03->10/31:04'
   },
   'europe/rome': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/samara': {
     offset: 4,
@@ -1775,12 +1775,12 @@ module.exports = {
   'europe/san_marino': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/sarajevo': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/simferopol': {
     offset: 3,
@@ -1789,32 +1789,32 @@ module.exports = {
   'europe/skopje': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/sofia': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:03->10/25:04'
+    dst: '03/28:03->10/31:04'
   },
   'europe/stockholm': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/tallinn': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:03->10/25:04'
+    dst: '03/28:03->10/31:04'
   },
   'europe/tirane': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/tiraspol': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/ulyanovsk': {
     offset: 4,
@@ -1823,27 +1823,27 @@ module.exports = {
   'europe/uzhgorod': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:03->10/25:04'
+    dst: '03/28:03->10/31:04'
   },
   'europe/vaduz': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/vatican': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/vienna': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/vilnius': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:03->10/25:04'
+    dst: '03/28:03->10/31:04'
   },
   'europe/volgograd': {
     offset: 4,
@@ -1856,22 +1856,22 @@ module.exports = {
   'europe/warsaw': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/zagreb': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'europe/zaporozhye': {
     offset: 3,
     hem: 'n',
-    dst: '03/29:03->10/25:04'
+    dst: '03/28:03->10/31:04'
   },
   'europe/zurich': {
     offset: 2,
     hem: 'n',
-    dst: '03/29:02->10/25:03'
+    dst: '03/28:02->10/31:03'
   },
   'indian/antananarivo': {
     offset: 3,
@@ -1920,27 +1920,27 @@ module.exports = {
   'mexico/bajanorte': {
     offset: -7,
     hem: 'n',
-    dst: '03/08:02->11/01:02'
+    dst: '03/14:02->11/07:02'
   },
   'mexico/bajasur': {
     offset: -6,
     hem: 'n',
-    dst: '04/05:02->10/25:02'
+    dst: '04/04:02->10/31:02'
   },
   'mexico/general': {
     offset: -5,
     hem: 'n',
-    dst: '04/05:02->10/25:02'
+    dst: '04/04:02->10/31:02'
   },
   'pacific/apia': {
     offset: 13,
     hem: 's',
-    dst: '04/05:04->09/27:03'
+    dst: '04/04:04->09/26:03'
   },
   'pacific/auckland': {
     offset: 12,
     hem: 's',
-    dst: '04/05:03->09/27:02'
+    dst: '04/04:03->09/26:02'
   },
   'pacific/bougainville': {
     offset: 11,
@@ -1949,7 +1949,7 @@ module.exports = {
   'pacific/chatham': {
     offset: 12.75,
     hem: 's',
-    dst: '04/05:03->04/05:02'
+    dst: '04/04:03->04/04:02'
   },
   'pacific/chuuk': {
     offset: 10,
@@ -1958,7 +1958,7 @@ module.exports = {
   'pacific/easter': {
     offset: -6,
     hem: 's',
-    dst: '04/04:22->09/05:22'
+    dst: '04/03:22->09/04:22'
   },
   'pacific/efate': {
     offset: 11,
@@ -1975,7 +1975,7 @@ module.exports = {
   'pacific/fiji': {
     offset: 12,
     hem: 's',
-    dst: '01/12:03->12/20:02'
+    dst: '01/17:03->11/14:02'
   },
   'pacific/funafuti': {
     offset: 12,
@@ -2040,7 +2040,7 @@ module.exports = {
   'pacific/norfolk': {
     offset: 11.5,
     hem: 'n',
-    dst: '04/05:03->10/04:02'
+    dst: '04/04:03->10/03:02'
   },
   'pacific/noumea': {
     offset: 11,